refactor(plots): plot-registry.tsx を責務別に階層分割
肥大化した単一ファイル(1365行)を plot-registry/ ディレクトリへ分割。
状態・データ取得・ハンドラは index.tsx に集約し、表示は単一責務の
子コンポーネントへ委譲する。挙動・マークアップ・testid は不変。

- index.tsx        : state/fetch/handlers + レイアウト合成
- PlotSearchBar    : 検索バー + 検索履歴ドロップダウン + 新規登録
- PlotToolbar      : フィルタ/あいう順トグル・条件クリア・列幅リセット・埋葬者表示
- PlotFilters      : 折りたたみフィルタ行
- PlotAiueoTabs    : あいう順タブ（キーボードナビ）
- PlotCardList     : モバイルカード一覧（md未満）
- PlotTable        : デスクトップテーブル（thead/tbody, 列幅・全文表示）
- PlotPagination   : ページネーション
- ColumnResizer    : 列幅ドラッグハンドル
- SortIndicator    : ソート方向インジケーター
- constants.ts     : AIUEO_TABS / *_LABELS / *_VARIANTS / SERVER_SORT_MAP
- utils.tsx        : 検索履歴 / フォーマッタ / getStatusBadge / getRowBgColor
- types.ts         : SortKey / SortOrder / AiueoTab / PlotRegistryProps

import パス '@/components/plot-registry' はディレクトリ index に解決されるため
呼び出し側(plots/page.tsx)は変更不要。

検証: lint 0 / tsc 0 / unit 345 passed / build 成功。

Closes #116
